feat: add use hooks
